# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IMachine": {
				"properties": {
					"id": {
						"type": "string"
					},
					"ip": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"addedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"ip",
					"name",
					"addedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IMachine.Exclude_keyofIMachine.id-or-addedAt__": {
				"properties": {
					"ip": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"ip",
					"name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IMachine.id-or-addedAt_": {
				"$ref": "#/components/schemas/Pick_IMachine.Exclude_keyofIMachine.id-or-addedAt__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"MachineCreationParam": {
				"$ref": "#/components/schemas/Omit_IMachine.id-or-addedAt_"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "marcys-link",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Joao Koritar ",
			"email": "jakoritarleite@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/machines/{machineId}": {
			"get": {
				"operationId": "GetMachine",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMachine"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "ip",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/machines": {
			"post": {
				"operationId": "CreateMachine",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MachineCreationParam"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}